Moved to a system of enemies having ID numbers, and we load the sprites for these enemies and generate them by ids. Lots of changes of counters to size_t.

+++ src/area.h
@@ -15,7 +15,7 @@ typedef struct Tilesheet {
 typedef struct Area {
     char *tilesetName;
     Tilesheet tilesheet;
-    int numRooms;
+    size_t numRooms;
 	Room *currentRoom;
     Room **roomList;
     char **entitySpriteNames;
